requirements.txt
asgiref==3.5.2
Django>=3
django-cors-headers>=3.0.0
djangorestframework>=3.0.0
gunicorn>=20.0.0
pytz==2022.1
sqlparse==0.4.2
whitenoise>=6.0.0
dj-database-url>=1.0.0
psycopg2-binary>=2.0.0
Pillow>=9.0.0

